requirements.txt
python-dotenv==1.0.0
requests==2.31.0
supabase==2.28.0
schedule==1.2.2
